Tools: Refactor: decouple version info with version checker
