+++ .github/workflows/package-index-sync.yaml
@@ -79,7 +79,7 @@ jobs:
           echo "Generate Only" > $RUNNER_TEMP/type.txt
           export CI_RUN_ID=${{ github.run_id }}
           export CI_RUN_URL=${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
-          python assets/renew_ruyi_index.py -c assets/config.toml -p . -i $RUNNER_TEMP/cache --log $RUNNER_TEMP/log.txt --warn $RUNNER_TEMP/warn.txt
+          python assets/renew_ruyi_index.py -p . -i $RUNNER_TEMP/cache --log $RUNNER_TEMP/log.txt --warn $RUNNER_TEMP/warn.txt
       - name: Run tool to generate and upload and PR
         if: ${{ ( github.event_name == 'workflow_dispatch' && inputs.makepr == 'true' ) || ( github.event_name == 'push' ) }}
         run: |
@@ -87,9 +87,9 @@ jobs:
           export CI_RUN_ID=${{ github.run_id }}
           export CI_RUN_URL=${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ github.run_id }}
           if ${{ ( github.event_name == 'workflow_dispatch' && inputs.force ) }}; then
-            python assets/renew_ruyi_index.py -c assets/config.toml -p . -i $RUNNER_TEMP/cache --log $RUNNER_TEMP/log.txt --warn $RUNNER_TEMP/warn.txt --pr --force
+            python assets/renew_ruyi_index.py -p . -i $RUNNER_TEMP/cache --log $RUNNER_TEMP/log.txt --warn $RUNNER_TEMP/warn.txt --pr --force
           else
-            python assets/renew_ruyi_index.py -c assets/config.toml -p . -i $RUNNER_TEMP/cache --log $RUNNER_TEMP/log.txt --warn $RUNNER_TEMP/warn.txt --pr
+            python assets/renew_ruyi_index.py -p . -i $RUNNER_TEMP/cache --log $RUNNER_TEMP/log.txt --warn $RUNNER_TEMP/warn.txt --pr
           fi
       - name: Output Summary
         run: |
