+++ assets/src/matrix_parser/__init__.py
@@ -3,3 +3,4 @@ Package for parsing metadata of the support-matrix
 """
 
 from .matrix_parser import *
+from .version import *
