+++ assets/src/ruyi_index_updator/upload_plugin/prelude.py
@@ -8,7 +8,7 @@ logger = logging.getLogger(__name__)
 
 try:
     from src.ruyi_index_updator.ruyi_index_parser import *
-    from src.version_checker import VInfo
+    from src.matrix_parser import VInfo
     from src.ruyi_index_updator.upload_plugin_base import UploadPluginBase
 except ModuleNotFoundError:
     # Things above is used for type hinting
