+++ assets/src/version_checker/__init__.py
@@ -1,3 +1,2 @@
-from .gen_oldver import gen_oldver, vinfo_dict_to_dict, VInfo
 from .run_nvchecker import run_nvchecker
 from .version_cmp import is_newer, is_older, is_same, filter_newer
